Fixed overflow for 32 bits (#825)

+++ pumpkin-util/src/math/mod.rs
@@ -49,7 +49,7 @@ pub const fn ceil_log2(value: u32) -> u8 {
         smallest_encompassing_power_of_two(value)
     };
 
-    MULTIPLY_DE_BRUIJN_BIT_POSITION[(((value as usize) * 125613361) >> 27) & 31]
+    MULTIPLY_DE_BRUIJN_BIT_POSITION[((((value as u64) * 125613361) >> 27) & 31) as usize]
 }
 
 /// Maximum return value: 30
